Bugfixes to get RT + DE (equilibrium) working. Yeah

+++ include/get_dust_thermal_emission.h
@@ -9,6 +9,6 @@
 //**********************************************************************
 // external function definitions
 extern void ComputeDustEmission (vector <float> & J, GrainModel & GrainModel, 
-				 vector <vector<float> > & EmmittedEnergy, bool DoStochastic);
+				 vector <vector<double> > & EmmittedEnergy, bool DoStochastic);
 
 #endif
